Refactor shared CodeTransform util functions. (#4255)
