+++ plugins/amazonq/codetransform/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codemodernizer/CodeModernizerStartupActivity.kt
@@ -5,7 +5,7 @@ package software.aws.toolkits.jetbrains.services.codemodernizer
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
-import software.aws.toolkits.jetbrains.services.codemodernizer.auth.isCodeTransformAvailable
+import software.aws.toolkits.jetbrains.services.codemodernizer.utils.isCodeTransformAvailable
 
 class CodeModernizerStartupActivity : StartupActivity.DumbAware {
