+++ plugins/amazonq/codetransform/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codemodernizer/utils/CodeTransformValidationUtils.kt
@@ -3,15 +3,9 @@
 
 package software.aws.toolkits.jetbrains.services.codemodernizer.utils
 
-import com.intellij.openapi.application.ApplicationInfo
 import com.intellij.openapi.project.Project
 import org.jetbrains.idea.maven.project.MavenProjectsManager
 import org.jetbrains.plugins.gradle.settings.GradleSettings
-
-fun isIntellij(): Boolean {
-    val productCode = ApplicationInfo.getInstance().build.productCode
-    return productCode == "IC" || productCode == "IU"
-}
 
 fun isGradleProject(project: Project) = !GradleSettings.getInstance(project).linkedProjectsSettings.isEmpty()
